+++ api/client/client_test.go
@@ -29,6 +29,12 @@ func (c mockAPIClient) Deploy(ctx context.Context, in *pb.DeployRequest,
 	return &pb.DeployReply{}, nil
 }
 
+func (c mockAPIClient) QueryCounters(ctx context.Context, in *pb.CountersRequest,
+	opts ...grpc.CallOption) (*pb.CountersReply, error) {
+
+	return &pb.CountersReply{}, nil
+}
+
 func (c mockAPIClient) Version(ctx context.Context, in *pb.VersionRequest,
 	opts ...grpc.CallOption) (*pb.VersionReply, error) {